+++ package.json
@@ -23,5 +23,8 @@
   "bugs": {
     "url": "https://github.com/vpratnicki/project-3/issues"
   },
-  "homepage": "https://github.com/vpratnicki/project-3#readme"
+  "homepage": "https://github.com/vpratnicki/project-3#readme",
+  "dependencies": {
+    "jwt-decode": "^3.1.2"
+  }
 }
